+++ src/components/Snackbar/SnackbarWrapper.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { SnackbarProps, useSnackbar } from './Snackbar2';
+import { SnackbarProps, useSnackbar } from './Snackbar';
 
 const SnackbarWrapper: React.FC<SnackbarProps> = props => {
   const { myEnqueueSnackbar } = useSnackbar();
