+++ src/huexports.ts
@@ -13,4 +13,4 @@ export { default as HuTabs } from './components/Tabs/Tabs';
 export { default as HuInputClassic } from './components/Input/InputClassic';
 export { default as HuFormInputClassic } from './components/Input/FormInputClassic';
 export { default as HuFormSelectionCard } from './components/SelectionCard/FormSelectionCard';
-export { default as HuUseSnackbar } from './components/SnackbarHugo/Snackbar2';
+export { default as HuUseSnackbar } from './components/Snackbar/Snackbar';
